+++ src/node/node_state.h
@@ -26,7 +26,6 @@
 #include "node/http_node_client.h"
 #include "node/jwt_key_auto_refresh.h"
 #include "node/node_to_node_channel_manager.h"
-#include "node/reconfig_id.h"
 #include "node/rpc/serdes.h"
 #include "node_to_node.h"
 #include "resharing.h"
@@ -73,7 +72,7 @@ namespace ccf
   struct NodeCreateInfo
   {
     crypto::Pem self_signed_node_cert;
-    crypto::Pem network_cert;
+    crypto::Pem service_cert;
   };
 
   void reset_data(std::vector<uint8_t>& data)
@@ -320,7 +319,9 @@ namespace ccf
         case StartType::Start:
         {
           network.identity = std::make_unique<ReplicatedNetworkIdentity>(
-            "CN=CCF Network", curve_id);
+            curve_id,
+            config.startup_host_time,
+            config.initial_service_certificate_validity_days);
 
           network.ledger_secrets->init();
 
@@ -379,7 +380,9 @@ namespace ccf
         case StartType::Recover:
         {
           network.identity = std::make_unique<ReplicatedNetworkIdentity>(
-            "CN=CCF Network", curve_id);
+            curve_id,
+            config.startup_host_time,
+            config.initial_service_certificate_validity_days);
 
           bool from_snapshot = !config.startup_snapshot.empty();
           setup_recovery_hook();
@@ -408,9 +411,13 @@ namespace ccf
     //
     void initiate_join()
     {
-      auto network_ca = std::make_shared<tls::CA>(config.join.network_cert);
+      auto network_ca = std::make_shared<tls::CA>(config.join.service_cert);
       auto join_client_cert = std::make_unique<tls::Cert>(
-        network_ca, self_signed_node_cert, node_sign_kp->private_key_pem());
+        network_ca,
+        self_signed_node_cert,
+        node_sign_kp->private_key_pem(),
+        tls::Auth::auth_required,
+        config.join.target_rpc_address);
 
       // Create RPC client and connect to remote node
       auto join_client =
@@ -1070,7 +1077,6 @@ namespace ccf
             "Could not commit transaction when finishing network recovery");
         }
       }
-      open_user_frontend();
       reset_data(quote_info.quote);
       reset_data(quote_info.endorsements);
       sm.advance(State::partOfNetwork);
@@ -1236,10 +1242,7 @@ namespace ccf
         }
 
         GenesisGenerator g(network, tx);
-        if (g.open_service())
-        {
-          open_user_frontend();
-        }
+        g.open_service();
         return;
       }
       else
@@ -1609,7 +1612,7 @@ namespace ccf
         create_params.node_endorsed_certificate);
 
       create_params.public_key = node_sign_kp->public_key_pem();
-      create_params.network_cert = network.identity->cert;
+      create_params.service_cert = network.identity->cert;
       create_params.quote_info = quote_info;
       create_params.public_encryption_key = node_encrypt_kp->public_key_pem();
       create_params.code_digest = node_code_id;
@@ -1819,8 +1822,20 @@ namespace ccf
               accept_network_tls_connections();
 
               open_frontend(ActorsType::members);
-              open_user_frontend();
             }
+          }));
+
+      network.tables->set_global_hook(
+        network.service.get_name(),
+        network.service.wrap_commit_hook(
+          [this](kv::Version hook_version, const Service::Write& w) {
+            if (!w.has_value())
+            {
+              throw std::logic_error("Unexpected deletion in service value");
+            }
+
+            network.identity->set_certificate(w->cert);
+            open_user_frontend();
           }));
     }
 
@@ -1979,14 +1994,6 @@ namespace ccf
           [](kv::Version version, const Nodes::Write& w)
             -> kv::ConsensusHookPtr {
             return std::make_unique<ConfigurationChangeHook>(version, w);
-          }));
-
-      network.tables->set_map_hook(
-        network.network_configurations.get_name(),
-        network.network_configurations.wrap_map_hook(
-          [](kv::Version version, const NetworkConfigurations::Write& w)
-            -> kv::ConsensusHookPtr {
-            return std::make_unique<NetworkConfigurationsHook>(version, w);
           }));
 
       network.tables->set_map_hook(